Non-matching decomp of func_800EB534_FA4E4

+++ include/functions.us.h
@@ -25,7 +25,7 @@ void func_800020E0_2CE0(s32 arg0, s32 arg1);
 void func_800035D8_41D8(s16 arg0);
 
 // F9230
-void func_800EB534_FA4E4(void *arg0, s32 arg1, s32 arg2, f32 arg3);
+void func_800EB534_FA4E4(void *arg0, s32 arg1, s32 arg2, s32 arg3);
 s32 func_800F066C_FF61C();
 void func_80000D0C_190C(void);
 s32 func_8000091C_151C(BhAudioTask *arg0, BhAudioBuf *arg1);
